Move UUID from change to general framework namespace
