+++ bundles/framework/tools.vitruv.framework.change.echange/src-gen/tools/vitruv/framework/change/echange/eobject/impl/EobjectPackageImpl.java
@@ -48,7 +48,7 @@ import tools.vitruv.framework.change.echange.root.RootPackage;
 
 import tools.vitruv.framework.change.echange.root.impl.RootPackageImpl;
 
-import tools.vitruv.framework.change.uuid.UuidPackage;
+import tools.vitruv.framework.uuid.UuidPackage;
 
 /**
  * <!-- begin-user-doc -->
