+++ bundles/framework/tools.vitruv.framework.change.echange/src-gen/tools/vitruv/framework/change/echange/impl/EChangePackageImpl.java
@@ -45,7 +45,7 @@ import tools.vitruv.framework.change.echange.root.RootPackage;
 
 import tools.vitruv.framework.change.echange.root.impl.RootPackageImpl;
 
-import tools.vitruv.framework.change.uuid.UuidPackage;
+import tools.vitruv.framework.uuid.UuidPackage;
 
 /**
  * <!-- begin-user-doc -->
