+++ bundles/framework/tools.vitruv.framework.change.echange/src/tools/vitruv/framework/change/echange/TypeInferringUnresolvingCompoundEChangeFactory.xtend
@@ -1,6 +1,6 @@
 package tools.vitruv.framework.change.echange
 
-import tools.vitruv.framework.change.uuid.UuidGeneratorAndResolver
+import tools.vitruv.framework.uuid.UuidGeneratorAndResolver
 
 final class TypeInferringUnresolvingCompoundEChangeFactory extends TypeInferringCompoundEChangeFactory {
