+++ bundles/framework/tools.vitruv.framework.change.uuid/META-INF/MANIFEST.MF
@@ -7,9 +7,9 @@ Bundle-ClassPath: .
 Bundle-Vendor: %providerName
 Bundle-Localization: plugin
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Export-Package: tools.vitruv.framework.change.uuid,
- tools.vitruv.framework.change.uuid.impl,
- tools.vitruv.framework.change.uuid.util
+Export-Package: tools.vitruv.framework.uuid.impl,
+ tools.vitruv.framework.uuid.util,
+ tools.vitruv.framework.uuid
 Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.emf.ecore;visibility:=reexport,
  com.google.guava,
