--- bundles/framework/tools.vitruv.framework.change.uuid/src-gen/tools/vitruv/framework/change/uuid/impl/UuidFactoryImpl.java
@@ -1,168 +1,0 @@
-/**
- */
-package tools.vitruv.framework.change.uuid.impl;
-
-import java.util.Map;
-
-import org.eclipse.emf.ecore.EClass;
-import org.eclipse.emf.ecore.EDataType;
-import org.eclipse.emf.ecore.EObject;
-import org.eclipse.emf.ecore.EPackage;
-
-import org.eclipse.emf.ecore.impl.EFactoryImpl;
-
-import org.eclipse.emf.ecore.plugin.EcorePlugin;
-
-import tools.vitruv.framework.change.uuid.*;
-
-/**
- * <!-- begin-user-doc -->
- * An implementation of the model <b>Factory</b>.
- * <!-- end-user-doc -->
- * @generated
- */
-public class UuidFactoryImpl extends EFactoryImpl implements UuidFactory {
-	/**
-	 * Creates the default factory implementation.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public static UuidFactory init() {
-		try {
-			UuidFactory theUuidFactory = (UuidFactory)EPackage.Registry.INSTANCE.getEFactory(UuidPackage.eNS_URI);
-			if (theUuidFactory != null) {
-				return theUuidFactory;
-			}
-		}
-		catch (Exception exception) {
-			EcorePlugin.INSTANCE.log(exception);
-		}
-		return new UuidFactoryImpl();
-	}
-
-	/**
-	 * Creates an instance of the factory.
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public UuidFactoryImpl() {
-		super();
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	@Override
-	public EObject create(EClass eClass) {
-		switch (eClass.getClassifierID()) {
-			case UuidPackage.UUID_TO_EOBJECT_REPOSITORY: return createUuidToEObjectRepository();
-			case UuidPackage.UUID_TO_EOBJECT_MAP_ENTRY: return (EObject)createUuidToEObjectMapEntry();
-			case UuidPackage.EOBJECT_TO_UUID_MAP_ENTRY: return (EObject)createEObjectToUuidMapEntry();
-			default:
-				throw new IllegalArgumentException("The class '" + eClass.getName() + "' is not a valid classifier");
-		}
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	@Override
-	public Object createFromString(EDataType eDataType, String initialValue) {
-		switch (eDataType.getClassifierID()) {
-			case UuidPackage.UUID:
-				return createUuidFromString(eDataType, initialValue);
-			default:
-				throw new IllegalArgumentException("The datatype '" + eDataType.getName() + "' is not a valid classifier");
-		}
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	@Override
-	public String convertToString(EDataType eDataType, Object instanceValue) {
-		switch (eDataType.getClassifierID()) {
-			case UuidPackage.UUID:
-				return convertUuidToString(eDataType, instanceValue);
-			default:
-				throw new IllegalArgumentException("The datatype '" + eDataType.getName() + "' is not a valid classifier");
-		}
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public UuidToEObjectRepository createUuidToEObjectRepository() {
-		UuidToEObjectRepositoryImpl uuidToEObjectRepository = new UuidToEObjectRepositoryImpl();
-		return uuidToEObjectRepository;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public Map.Entry<String, EObject> createUuidToEObjectMapEntry() {
-		UuidToEObjectMapEntryImpl uuidToEObjectMapEntry = new UuidToEObjectMapEntryImpl();
-		return uuidToEObjectMapEntry;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public Map.Entry<EObject, String> createEObjectToUuidMapEntry() {
-		EObjectToUuidMapEntryImpl eObjectToUuidMapEntry = new EObjectToUuidMapEntryImpl();
-		return eObjectToUuidMapEntry;
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public String createUuidFromString(EDataType eDataType, String initialValue) {
-		return (String)super.createFromString(eDataType, initialValue);
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public String convertUuidToString(EDataType eDataType, Object instanceValue) {
-		return super.convertToString(eDataType, instanceValue);
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @generated
-	 */
-	public UuidPackage getUuidPackage() {
-		return (UuidPackage)getEPackage();
-	}
-
-	/**
-	 * <!-- begin-user-doc -->
-	 * <!-- end-user-doc -->
-	 * @deprecated
-	 * @generated
-	 */
-	@Deprecated
-	public static UuidPackage getPackage() {
-		return UuidPackage.eINSTANCE;
-	}
-
-} //UuidFactoryImpl
