+++ bundles/framework/tools.vitruv.framework.change/src/tools/vitruv/framework/change/description/impl/ConcreteApplicableChangeImpl.xtend
@@ -4,7 +4,7 @@ import org.eclipse.emf.ecore.EObject
 import tools.vitruv.framework.change.echange.EChange
 import tools.vitruv.framework.tuid.TuidManager
 import static extension tools.vitruv.framework.change.echange.resolve.EChangeResolverAndApplicator.*
-import tools.vitruv.framework.change.uuid.UuidResolver
+import tools.vitruv.framework.uuid.UuidResolver
 import tools.vitruv.framework.change.echange.resolve.EChangeUnresolver
 
 class ConcreteApplicableChangeImpl extends ConcreteChangeImpl {
