+++ bundles/framework/tools.vitruv.framework.vsum/src/tools/vitruv/framework/vsum/repositories/ResourceRepositoryImpl.java
@@ -24,9 +24,9 @@ import org.eclipse.emf.transaction.TransactionalEditingDomain;
 import edu.kit.ipd.sdq.commons.util.org.eclipse.emf.common.util.URIUtil;
 import tools.vitruv.framework.change.description.TransactionalChange;
 import tools.vitruv.framework.change.recording.AtomicEmfChangeRecorder;
-import tools.vitruv.framework.change.uuid.UuidGeneratorAndResolver;
-import tools.vitruv.framework.change.uuid.UuidGeneratorAndResolverImpl;
-import tools.vitruv.framework.change.uuid.UuidResolver;
+import tools.vitruv.framework.uuid.UuidGeneratorAndResolver;
+import tools.vitruv.framework.uuid.UuidGeneratorAndResolverImpl;
+import tools.vitruv.framework.uuid.UuidResolver;
 import tools.vitruv.framework.correspondence.CorrespondenceModel;
 import tools.vitruv.framework.correspondence.CorrespondenceModelImpl;
 import tools.vitruv.framework.correspondence.CorrespondenceProviding;
