+++ tests/framework/tools.vitruv.framework.change.tests/src/tools/vitruv/framework/tests/change/ChangeDescription2ChangeTransformationTest.xtend
@@ -16,9 +16,9 @@ import org.eclipse.emf.ecore.xmi.impl.XMIResourceFactoryImpl
 import java.util.ArrayList
 import tools.vitruv.framework.change.recording.AtomicEmfChangeRecorder
 import tools.vitruv.framework.util.bridges.EMFBridge
-import tools.vitruv.framework.change.uuid.UuidGeneratorAndResolverImpl
+import tools.vitruv.framework.uuid.UuidGeneratorAndResolverImpl
 import static extension tools.vitruv.framework.change.echange.resolve.EChangeResolverAndApplicator.*;
-import tools.vitruv.framework.change.uuid.UuidGeneratorAndResolver
+import tools.vitruv.framework.uuid.UuidGeneratorAndResolver
 import tools.vitruv.framework.change.echange.resolve.EChangeUnresolver
 
 /**
